backend/lct_backend1/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
Pillow==10.1.0
aiofiles==23.2.1
boto3==1.34.0
aioboto3
aio-pika
aiormq
